+++ src/OpenTelemetry.Resources.Process/ProcessDetector.cs
@@ -11,7 +11,7 @@ namespace OpenTelemetry.Resources.Process;
 /// </summary>
 internal sealed class ProcessDetector : IResourceDetector
 {
-    private static readonly Version SemanticConventionsVersion = new(1, 43, 0);
+    private static readonly Version SemanticConventionsVersion = new(1, 44, 0);
 
     /// <summary>
     ///     Detects the resource attributes for process.