+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.19.1`.
+  ([#5329](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5329))
+
+## 0.3.0-alpha.2
+
+Released 2026-Sep-18
+
 * Fixed Kafka spans exporting unsanitized exception messages as status descriptions.
   ([#5214](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5214))
+
+* Updated OpenTelemetry core component version(s) to `1.19.0`.
+  ([#5240](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5240))
 
 ## 0.3.0-alpha.1